example/slack/pyproject.toml
[tool.poetry]
name = "slack"
version = "0.1.0"
description = ""
authors = ["MarkInada <makotoinada132@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.1"
slack-bolt = "^1.18.0"
openai = "^0.28.0"
langchain = "^0.0.281"
stateful-chat-client = "^0.1.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
